omer : removed ts from backend
