--- backend/nodemon.json
@@ -1,7 +1,0 @@
-{
-    "watch": [
-        ""
-    ],
-    "ext": "ts",
-    "exec": "./node_modules/.bin/ts-node-esm index.ts"
-}
