+++ backend/package.json
@@ -2,13 +2,11 @@
   "name": "backend",
   "version": "1.0.0",
   "description": "",
-  "main": "index.ts",
+  "main": "index.js",
   "type": "module",
   "scripts": {
     "dev": "nodemon",
-    "postinstall": "npm run build",
-    "start": "node build/index.js",
-    "build": "tsc"
+    "start": "node index.js"
   },
   "author": "",
   "license": "ISC",
@@ -17,12 +15,9 @@
     "cors": "^2.8.5",
     "dotenv": "^16.0.2",
     "expess": "^0.0.1-security",
-    "express": "^4.18.1",
-    "ts-node": "^10.9.1",
-    "typescript": "^4.8.2"
+    "express": "^4.18.1"
   },
   "devDependencies": {
-    "@types/express": "^4.17.13",
     "nodemon": "^2.0.19"
   }
 }
