--- backend/tsconfig.json
@@ -1,9 +1,0 @@
-{
-    "compilerOptions": {
-        "target": "ES6",
-        "allowSyntheticDefaultImports": true,
-        "moduleResolution": "node",
-        "module": "esnext",
-        "outDir": "build"
-    }
-}
